+++ examples/getting-started/section7/src/cjson/cJSON.h
@@ -1,0 +1,4 @@
+#ifndef CJSON_H
+#define CJSON_H
+int cjson_parse_int(const char *s);
+#endif